+++ src/index.js
@@ -1,10 +1,10 @@
-import BreakpointContainer, { BrowserContainer } from './components/BreakpointContainer';
+import { BreakpointContainer, BrowserContainer } from './components/BreakpointContainer';
 import { withBreakpointContainer, withBrowserContainer } from './components/HOCs';
 import { Breakpoint } from './components/Breakpoint';
 import { BREAKPOINTS, getBpUpperLimit } from './data/breakpoints';
 
-export default BreakpointContainer;
 export {
+	BreakpointContainer,
 	BrowserContainer,
 	withBreakpointContainer,
 	withBrowserContainer,